+++ src/postorius/doc/news.rst
@@ -28,6 +28,7 @@ UI
 
 * Add 'Delivery Mode' and 'Moderation Action' columns to List members
   page. (See #471)
+* Add support to list and handle pending un-subscription requests. (Closes #332)
 
 Other
 -----
@@ -54,7 +55,7 @@ Other
 * Show held message is local time of the User. (Closes #434)
 * Fix a bug where non-member options page would show an owner's options if the
   same email was subscribed as owner and non-member in a list. (Closes #436)
-* Switching subscription from one email address to other or Primary Address now 
+* Switching subscription from one email address to other or Primary Address now
   preserves preferences and does not require Moderator approval. (Closes #425)
 * Make 'Archives' and 'List Options' urls more prominently visible in the
   list summary page as buttons. (Closes #439)